+++ Demo/IMOStyledTableViewControllerDemo/ICSheetTableViewController.m
@@ -34,12 +34,12 @@
 }
 
 
-- (int)numberOfSectionsInTableView:(UITableView *)tableView {
+- (NSInteger)numberOfSectionsInTableView:(UITableView *)tableView {
     return 2;
 }
 
 
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
     int rows = 0;
     switch (section) {
         case 0:
